# Replace TrimSpace to TrimRightFunc