--- a/AUTHORS
+++ b/AUTHORS
@@ -7,3 +7,6 @@ They listed below in an alphabetical order:
 - Kz26
 - Makombo
 - Zac Shenker <zshenker@brightcove.com>
+
+If you want to be added to this list (or removed for any reason)
+just open issue about it.