--- a/AUTHORS
+++ b/AUTHORS
@@ -2,6 +2,7 @@ There several persons contribute their code to the project.
 They listed below in an alphabetical order:
 
 - Alexander (Axel) I.Grafov <grafov@gmail.com>
+- Denys Smirnov <denis.smirnov.91@gmail.com>
 - Fabrizio (Misto) Milo <mistobaan@gmail.com>
 - Jamie Stackhouse <jamie.stackhouse@redspace.com>
 - Kz26