--- a/AUTHORS
+++ b/AUTHORS
@@ -1,12 +1,15 @@
-There are several persons contribute their code to the project.
-They listed below in an alphabetical order:
+There are several persons contribute their code include small patches
+to the project. They listed below in an alphabetical order:
 
 - Alexander (Axel) I.Grafov <grafov@gmail.com>
+- Andrey Chernov <chernov@bradburylab.com>
 - Denys Smirnov <denis.smirnov.91@gmail.com>
 - Fabrizio (Misto) Milo <mistobaan@gmail.com>
 - Jamie Stackhouse <jamie.stackhouse@redspace.com>
+- Julian Cooper <jcooper@brightcove.com>
 - Kz26
 - Makombo
+- Scott Kidder <skidder@brightcove.com>
 - Zac Shenker <zshenker@brightcove.com>
 
 If you want to be added to this list (or removed for any reason)