--- a/AUTHORS
+++ b/AUTHORS
@@ -1,4 +1,4 @@
-There are many persons contribute their code include small patches
+There are many persons contribute their code (include small patches)
 to the project. They listed below in an alphabetical order:
 
 - Alexander (Axel) I.Grafov <grafov@gmail.com>
@@ -17,4 +17,4 @@ to the project. They listed below in an alphabetical order:
 - Zac Shenker <zshenker@brightcove.com>
 
 If you want to be added to this list (or removed for any reason)
-just open issue about it.
+just open an issue about it.