--- a/sample-playlists/media-playlist-with-discontinuity.m3u8
+++ b/sample-playlists/media-playlist-with-discontinuity.m3u8
@@ -1,0 +1,15 @@
+# https://developer.apple.com/library/ios/technotes/tn2288/_index.html
+#
+#EXTM3U
+#EXT-X-TARGETDURATION:10
+#EXT-X-VERSION:3
+#EXT-X-MEDIA-SEQUENCE:0
+#EXTINF:10.0,
+ad0.ts
+#EXTINF:8.0,
+ad1.ts
+#EXT-X-DISCONTINUITY
+#EXTINF:10.0,
+movieA.ts
+#EXTINF:10.0,
+movieB.ts